Melodious quest
Hi im the one and only Melodious Milli, 
Ive always been a huge fan of music and art and literature. Im currently learning to have bachlors in music and fine arts. Im slowly starting to understand how complex it is to not only write but as well as compose the songs that we listen to today. I would love to learn more and collaborate with a few artist on some ideas i have in mind that i think would be a sucess. You can contact me by email at melodymccall8282@gmail.com